+++ Basil/include/IndexBuffer.h
@@ -10,7 +10,7 @@ namespace Basil
 	class IndexBuffer
 	{
 		public:
-			virtual ~IndexBuffer() {};
+			virtual ~IndexBuffer() {}
 			virtual void bind() const = 0;
 			virtual void unbind() const = 0;
 			virtual uint32_t getSize() const = 0;